Replace interactive prompts with "expect" scripts
This patch adds a new dependency to the setup: expect. It is necessary
because this change replaces interaction with protected password inputs
which won't work with simple echos.

New helper scripts were based on "expect_autopasswd" script shipped
with "expect" package as well as "autoexpect" output for corresponding
commands.

Remember to change sample admin password before bringing up this setup.

Fixes: #2

Signed-off-by: Paweł Wieczorek <pawiecz@collabora.com>